Set the source track information in the ingress state (#121)
This gathers information either from the GStreamer stream, or the SDP to populate the audio and video state entries in the ingress state. This useful for debugging, analytics and billing.

This also refactors how state update RPCs are sent, delegating it to the Params object most of the time.

Ingress State updates in the Params object have also been made thread safe.