Add a media watchdog (#334)
This adds a media watchdog at the input of the go SDK (so at the tail of the pipeline). This is needed because in some cases, such as SRT input. gstreamer will retry to connect for ever without emitting any event about the failure to connect. This will close the ingress if no media is received on any track for 1 min.