Back down synchronizer wait time if the processing pipeline becomes too long (#337)
Currently, the output synchronizer increases the reference/wait time if a packet arrives later than expected at the output to compensate drifts, stalls or other processing hiccups. The reference time is however never brought back down if the clock drifts the other way, or the total pipeline processing time decreases. In the case of SRT where there is a guaranteed latency between client and server, this can cause the input buffer to fill up to its max value, causing essentially a deadlock.

This patch adds a queue we can monitor between the appsink and the go SDK. We measure the length of this queue and if it grows to more than 2 buffers, we decrease the synchronizer wait time in quantum of 10ms until the queue shortens below the threshold.

The above is disabled to HLS as for VOD (or live with a large buffer) content, the back pressure at the output is needed and the pipeline is expected to hold a lot of data.