Make sure EOS reaches sinks (#408)
* Make sure EOS reaches sinks
New sample callback (app sink callback) is doing push to a blocking queue and in case we are e.g blocked inside output writeSample - inside WaitForMediaTime , there is nothing to pump the data out and the app sink thread gets stuck on the blocking push, not being able to process EOS on time

The change singnals EOS seen on the source out of band - so the output gets a chance to decide to wait a bit for remaining data to go through but also to cancel and close if it takes too much. This should prevent pipeline frozen scenario (where EOS doesn't reach the sink). Eventually we can get rid of the existing workaround - but leaving it just as a safty net for now.

* tidy logging